pyproject.toml
[tool.poetry]
name = "cwscript-py"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "cwscript_py"}]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
antlr4-python3-runtime = "^4.12.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"